Split wrapper managed-bin installer tests
